requirements.txt
Django==3.1.3
djangorestframework==3.12.2
sqlparse==0.4.1
djangorestframework-jwt==1.11.0
Pillow==8.0.1
djangorestframework-files==1.1.0
django-upload-validator==1.1.5
